Move conductor into conductor directory
