+++ arangod/Pregel/Conductor/Conductor.cpp
@@ -37,14 +37,14 @@
 #include "Pregel/AggregatorHandler.h"
 #include "Pregel/AlgoRegistry.h"
 #include "Pregel/Algorithm.h"
-#include "Pregel/Conductor/CanceledState.h"
-#include "Pregel/Conductor/ComputingState.h"
-#include "Pregel/Conductor/DoneState.h"
-#include "Pregel/Conductor/FatalErrorState.h"
-#include "Pregel/Conductor/InErrorState.h"
-#include "Pregel/Conductor/LoadingState.h"
-#include "Pregel/Conductor/State.h"
-#include "Pregel/Conductor/StoringState.h"
+#include "Pregel/Conductor/States/CanceledState.h"
+#include "Pregel/Conductor/States/ComputingState.h"
+#include "Pregel/Conductor/States/DoneState.h"
+#include "Pregel/Conductor/States/FatalErrorState.h"
+#include "Pregel/Conductor/States/InErrorState.h"
+#include "Pregel/Conductor/States/LoadingState.h"
+#include "Pregel/Conductor/States/State.h"
+#include "Pregel/Conductor/States/StoringState.h"
 #include "Pregel/MasterContext.h"
 #include "Pregel/PregelFeature.h"
 #include "Pregel/Status/ConductorStatus.h"
