+++ arangod/Pregel/Conductor/Conductor.h
@@ -35,15 +35,15 @@
 #include "Utils/DatabaseGuard.h"
 
 #include "Pregel/ExecutionNumber.h"
-#include "Pregel/Conductor/CanceledState.h"
-#include "Pregel/Conductor/ComputingState.h"
-#include "Pregel/Conductor/DoneState.h"
-#include "Pregel/Conductor/FatalErrorState.h"
-#include "Pregel/Conductor/InErrorState.h"
-#include "Pregel/Conductor/InitialState.h"
-#include "Pregel/Conductor/LoadingState.h"
-#include "Pregel/Conductor/State.h"
-#include "Pregel/Conductor/StoringState.h"
+#include "Pregel/Conductor/States/CanceledState.h"
+#include "Pregel/Conductor/States/ComputingState.h"
+#include "Pregel/Conductor/States/DoneState.h"
+#include "Pregel/Conductor/States/FatalErrorState.h"
+#include "Pregel/Conductor/States/InErrorState.h"
+#include "Pregel/Conductor/States/InitialState.h"
+#include "Pregel/Conductor/States/LoadingState.h"
+#include "Pregel/Conductor/States/State.h"
+#include "Pregel/Conductor/States/StoringState.h"
 #include "Pregel/Status/ConductorStatus.h"
 #include "Pregel/Status/ExecutionStatus.h"
 #include "velocypack/Builder.h"
