+++ arangod/Pregel/Conductor/States/CMakeLists.txt
@@ -1,0 +1,9 @@
+target_sources(arango_pregel PRIVATE
+  InitialState.cpp
+  LoadingState.cpp
+  ComputingState.cpp
+  StoringState.cpp
+  CanceledState.cpp
+  DoneState.cpp
+  InErrorState.cpp
+  FatalErrorState.cpp)
