+++ arangod/Pregel/Conductor/States/ComputingState.h
@@ -22,9 +22,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
-#include <chrono>
-
-#include "Pregel/Conductor/State.h"
+#include "State.h"
 
 namespace arangodb::pregel {
 
@@ -32,19 +30,18 @@ class Conductor;
 
 namespace conductor {
 
-struct InError : State {
-  std::chrono::system_clock::time_point expiration;
+struct Computing : State {
   Conductor& conductor;
-  InError(Conductor& conductor, std::chrono::seconds const& ttl);
-  ~InError(){};
-  auto run() -> void override{};
+  Computing(Conductor& conductor);
+  ~Computing();
+  auto run() -> void override;
   auto receive(Message const& message) -> void override;
-  auto name() const -> std::string override { return "in error"; };
-  auto isRunning() const -> bool override { return false; }
+  auto name() const -> std::string override { return "running"; };
+  auto isRunning() const -> bool override { return true; }
   auto getExpiration() const
       -> std::optional<std::chrono::system_clock::time_point> override {
-    return expiration;
-  };
+    return std::nullopt;
+  }
 };
 
 }  // namespace conductor
