+++ arangod/Pregel/Conductor/States/DoneState.h
@@ -22,7 +22,9 @@
 ////////////////////////////////////////////////////////////////////////////////
 #pragma once
 
-#include "Pregel/Conductor/State.h"
+#include <chrono>
+
+#include "State.h"
 
 namespace arangodb::pregel {
 
@@ -30,18 +32,20 @@ class Conductor;
 
 namespace conductor {
 
-struct Storing : State {
+struct Done : State {
+  std::chrono::system_clock::time_point expiration;
   Conductor& conductor;
-  Storing(Conductor& conductor);
-  ~Storing();
+  Done(Conductor& conductor, std::chrono::seconds const& ttl);
+  ~Done() = default;
   auto run() -> void override;
   auto receive(Message const& message) -> void override;
-  auto name() const -> std::string override { return "storing"; };
-  auto isRunning() const -> bool override { return true; }
+  auto getResults(bool withId) -> PregelResults override;
+  auto name() const -> std::string override { return "done"; };
+  auto isRunning() const -> bool override { return false; }
   auto getExpiration() const
       -> std::optional<std::chrono::system_clock::time_point> override {
-    return std::nullopt;
-  }
+    return expiration;
+  };
 };
 
 }  // namespace conductor
