+++ arangod/Pregel/Conductor/States/InitialState.cpp
@@ -1,5 +1,5 @@
 #include "InitialState.h"
-#include "Pregel/Conductor.h"
+#include "Pregel/Conductor/Conductor.h"
 #include "Pregel/WorkerConductorMessages.h"
 
 using namespace arangodb::pregel::conductor;
