+++ arangod/Pregel/Conductor/States/StoringState.cpp
@@ -1,8 +1,7 @@
 #include "StoringState.h"
 
-#include "Pregel/Conductor.h"
+#include "Pregel/Conductor/Conductor.h"
 #include "Metrics/Gauge.h"
-#include "Pregel/Conductor/State.h"
 #include "Pregel/MasterContext.h"
 #include "Pregel/PregelFeature.h"
 #include "Pregel/WorkerConductorMessages.h"
